Use vmalloc rather than kzalloc for shadow/backing framebuffer.
This buffer is a copy of the framebuffer, used to scan for unchanged
pixels (so we can avoid sending them over usb, our big bottleneck).
The way we use it, there's no need for the pages to be physically contiguous
making vmalloc a good method, if we have this shadow/backing buffer at all.

Practically, kzalloc of this xMB allocation fails to on certain
configurations, most seen on RedHat and CentOS. vmalloc has no trouble.

Patch originally submitted by Dan Ellis, DisplayLink