Asynchronous URBs and refcounting for clean teardown
Incorporates recommended elements of usb-skel, including alloc/free of usb buffers dynamically.  Unfortunately, recent kernels have a WARN_ON which is triggered by freeing a buffer in a completion routine.
So following the usb-skel model is broken.  Will look to preallocate urbs and buffers to avoid this (future change).

Compiles, runs, but hits that WARN_ON on every transfer.