Fix greenscreen
Required a re-think of how the kernel mode driver could not touch framebuffer
memory after initializing defio. Cleaned up relationship between defio and
non-defio setup and codepaths.

non-defio mode is triggered by damage ioctl (it's a "special client" who
knows about the private ioctls). We save the effort of double-processing
changed pages and damage rect, but as it stands can't switch out the mmap,
so still take the hit of a page fault in the defio main code.