Support writing default/backup edid via /sys/class/graphics/fb?/edid
Required making udlfb much more flexible to let edid be dynamic
and changable, and then also to support changing/reallocating the
framebuffer size based on the (new) max mode supported by the monitor.

Writing a backup edid blob to the 'edid' file will cause udlfb to
attempt to re-read the edid from hardware.
If we still don't have a good edid, it will use the backup edid instead.

The intent is to 1) enable user-supplied edid
2) enable udev rules to backup the last known good edid
for a device, and then provide that edid as a backup for all future starts